+++ src/rest_api/classes/gene/widgets/interactions.clj
@@ -3,7 +3,7 @@
    [clojure.string :as str]
    [datomic.api :as d]
    [pseudoace.utils :as pace-utils]
-   [rest-api.classes.generic :as generic]
+   [rest-api.classes.generic-fields :as generic]
    [rest-api.formatters.object :refer [pack-obj]]))
 
 (def ^:private sort-by-id (partial sort-by :id))