+++ src/rest_api/classes/pcr_product.clj
@@ -1,0 +1,11 @@
+(ns rest-api.classes.pcr-product
+  (:require
+    ;[rest-api.classes.pcr-product.widgets.overview :as overview]
+    [rest-api.routing :as routing]))
+
+(routing/defroutes
+  {:entity-ns "pcr-product"
+   :uri-name "pcr-oligo"
+   :widget
+   {;:overview overview/widget
+    }})