+++ src/rest_api/classes/person/widgets/overview.clj
@@ -72,7 +72,13 @@
      :description "email addresses of this person"}))
 
 (defn also-known-as [person]
-  {:data (:person/also-known-as person)
+  {:data (when-let [names (:person/also-known-as person)]
+           (not-empty
+             (remove
+               nil?
+               (for [n names]
+                 (when (not= (:label (pack-obj person)) n)
+                   n)))))
    :description "other names person is also known as."})
 
 (defn previous-addresses [person]
@@ -87,7 +93,7 @@
                       (pace-utils/vmap
                         :date-modified (date/format-date (:person.old-address/datetype old-address))
                         :email (:address/email old-address)
-                        ;;                        :institution (:address/institution old-address)
+                        :institution (first (:address/institution old-address))
                         :street-address (:address/street-address old-address)
                         :country (:address/country old-address)
                         :main-phone (:address/main-phone old-address)
@@ -124,7 +130,7 @@
   {:name                     name-field
    :email                    email
    :orcid                    orcid
-   ;;    :institution              institution
+   :institution              institution
    :web_page                 web-page
    :street_address           street-address
    :aka                      also-known-as