+++ src/rest_api/db/sql/sequence.sql
@@ -1,5 +1,5 @@
--- :name gene-features :? :*
--- :doc Retrieve all sequences for a gene by id
+-- :name get-features :? :*
+-- :doc Retrieve all sequences for a feature by id
 SELECT f.id,f.object,f.typeid,t.tag as type,f.seqid,l.seqname,f.start,f.end,f.strand
 FROM feature as f
 JOIN name as n ON n.id=f.id
@@ -8,7 +8,7 @@ JOIN typelist as t ON t.id=f.typeid
 WHERE n.name = :name
 
 -- :name sequence-features-where-type :? :*
--- :doc Retrieve all sequences for a gene by id whre of type transcript, CDS, mRNA or scpecified type
+-- :doc Retrieve all sequences for a feautre by id whre of type transcript, CDS, mRNA or scpecified type
 SELECT f.id,f.object,f.typeid,t.tag as type,f.seqid,l.seqname,f.start,f.end,f.strand
 FROM feature as f
 JOIN name as n ON n.id=f.id