+++ src/rest_api/formatters/date.clj
@@ -31,3 +31,15 @@
    (.format
      (java.text.SimpleDateFormat. "dd MMM yyyy")
         date-str))
+
+(defn format-date5
+  [date-str]
+   (.format
+     (java.text.SimpleDateFormat. "yyyy")
+        date-str))
+
+(defn format-date6
+  [date-str]
+   (.format
+     (java.text.SimpleDateFormat. "dd MMM yyyy HH:mm:ss")
+        date-str))