+++ src/rest_api/formatters/object.clj
@@ -39,6 +39,14 @@
   (or (:gene/public-name obj)
       (:gene/id obj)))
 
+(defmethod obj-label "wbprocess" [_ obj]
+  (or (:wbprocess/public-name obj)
+      (:wbprocess/id obj)))
+
+(defmethod obj-label "laboratory" [_ obj]
+  (or (first (:laboratory/mail obj))
+      (:laboratory/id obj)))
+
 (defmethod obj-label "phenotype" [_ obj]
   (or (->> (:phenotype/primary-name obj)
            (:phenotype.primary-name/text))
@@ -80,11 +88,13 @@
      (str (author-lastname (first authors)) " et al."))))
 
 (defmethod obj-label "paper" [_ paper]
-  (if (seq (:paper/publication-date paper))
-    (str (author-list paper)
-         ", "
-         (first (str/split (:paper/publication-date paper)
-                           #"-")))))
+  (if-let [year (when (seq (:paper/publication-date paper))
+                  (first
+                    (str/split
+                      (:paper/publication-date paper)
+                      #"-")))]
+    (str (author-list paper) ", " year)
+    (author-list paper)))
 
 (defmethod obj-label "feature" [_ feature]
   (or (:feature/public-name feature)
@@ -100,11 +110,14 @@
   (:do-term/name term))
 
 (defmethod obj-label "person" [_ person]
-  (:person/standard-name person))
+ (or (:person/standard-name person)
+     (or (:author/id (first (:person/possibly-publishes-as person)))
+         (:person/id person))))
 
 (defmethod obj-label "construct" [_ cons]
   (or (first (:construct/public-name cons))
-      (:construct/id cons)))
+      (or (first (:construct/other-name cons))
+          (:construct/id cons))))
 
 (defmethod obj-label "transgene" [_ tg]
   (or (:transgene/public-name tg)
@@ -286,11 +299,9 @@
            :label pa}))
 
    :Author_evidence
-   (seq (for [a (:evidence/author-evidence holder)
-              :let [author (:evidence.author-evidence/author holder)]]
-          {:evidence
-           ;; Notes seem to be ignored here.
-           (pack-obj "author" author)}))
+   (seq (for [ah (:evidence/author-evidence holder)
+              :let [author (:evidence.author-evidence/author ah)]]
+          {:evidence (pack-obj "author" author)}))
 
    :Accession_evidence
    (if-let [accs (:evidence/accession-evidence holder)]
@@ -298,7 +309,7 @@
             db  :evidence.accession-evidence/database} accs]
        {:id acc
         :label (format "%s:%s" (:database/id db) acc)
-        :class (:database/id acc)}))
+        :class (:database/id db)}))
 
 
    :Protein_id_evidence
@@ -369,8 +380,8 @@
         (str/replace #"-" " ")
         (str/capitalize))))
 
-(defn name-field [gene]
-  (let [data (pack-obj gene)]
+(defn name-field [object]
+  (let [data (pack-obj object)]
     {:data (not-empty data)
      :description (format "The name and WormBase internal ID of %s"
                           (:id data))}))